requirements.txt
# Core Dependencies
anthropic==0.39.0
openai==1.54.0
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-openai==0.2.9
langchain-community==0.3.5

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Forensics & Security
cryptography==43.0.1
hashlib-additional==1.0.0

# Network Analysis
scapy==2.5.0
dpkt==1.9.8

# File Analysis
python-magic==0.4.27
# magic==5.32.1

# Database & Storage
sqlalchemy==2.0.35

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
jinja2==3.1.4
markupsafe==2.1.5

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Logging
loguru==0.7.2

# Date/Time
python-dateutil==2.9.0